requirements.txt
python-dotenv
aiogram
aiohttp
pytoniq
telebot
pytoncenter==0.0.14
# tonpy


